backend/requirements.txt
# ============================================
# StudyBuddy AI - Backend Dependencies
# ============================================

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Databases
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
motor==3.3.2
pymongo==4.6.1
redis==5.0.1
aiosqlite==0.19.0

# AI/ML - LangChain Ecosystem
langchain==0.1.0
langchain-core==0.1.10
langchain-community==0.0.13
langchain-google-genai==0.0.6
langgraph==0.0.20
google-generativeai==0.3.2

# Embeddings & Vector Store
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.26.3

# Content Processing - PDF
PyPDF2==3.0.1
pdfplumber==0.10.3
pypdf==3.17.4

# Content Processing - Documents  
python-docx==1.1.0
python-pptx==0.6.23

# Content Processing - Audio/Video
faster-whisper==0.10.0
pydub==0.25.1

# Content Processing - OCR
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.17.0

# Knowledge Graph
networkx==3.2.1

# NLP Tools
spacy==3.7.2

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
orjson==3.9.10
python-dateutil==2.8.2
uuid6==2024.1.12

# WebSockets
websockets==12.0

# Logging & Monitoring
structlog==24.1.0
python-json-logger==2.0.7
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0
factory-boy==3.3.0

# Development
black==24.1.0
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
